# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AR Qudrix ISP OS API",
    "description": "All-in-one Broadband ISP Management Platform with BTRC Compliance, Offline Support & Multi-Tenant Architecture",
    "version": "1.0.0",
    "contact": {
      "name": "AR Qudrix Support",
      "url": "https://arqudrix.com",
      "email": "support@arqudrix.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://arqudrix.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Development server"
    },
    {
      "url": "https://app.arqudrix.com/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/support/tickets": {
      "get": {
        "tags": ["Support"],
        "summary": "List support tickets",
        "operationId": "listTickets",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["open", "in-progress", "resolved", "closed"] }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": { "type": "string", "enum": ["low", "medium", "high", "urgent"] }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Ticket" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Support"],
        "summary": "Create support ticket",
        "operationId": "createTicket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TicketInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ticket" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
...
      "Ticket": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "subject": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["open", "in-progress", "resolved", "closed"] },
          "priority": { "type": "string", "enum": ["low", "medium", "high", "urgent"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "TicketInput": {
        "type": "object",
        "required": ["subject", "description"],
        "properties": {
          "subject": { "type": "string" },
          "description": { "type": "string" },
          "customerId": { "type": "string" },
          "priority": { "type": "string", "enum": ["low", "medium", "high", "urgent"] },
          "attachments": {
            "type": "array",
            "items": { "type": "string", "format": "binary" }
          }
        }
      },
...
    }
  }
}